# src/frontend/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ocelescope Backend",
    "version": "0.1"
  },
  "paths": {
...
    "/ocels/": {
      "get": {
        "tags": [
          "ocels"
        ],
        "summary": "List uploaded and uploading OCELs",
        "description": "Returns metadata for all uploaded OCELs along with any OCEL files currently being imported. Includes the ID of the currently active OCEL, if one is selected.",
        "operationId": "getOcels",
        "parameters": [
          {
            "name": "extension_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Extension Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OcelMetadata"
                  },
                  "title": "Response Getocels"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ocels"
        ],
        "summary": "Set Filter",
        "operationId": "setFilters",
        "parameters": [
          {
            "name": "ocel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ocel Id"
            }
          },
          {
            "name": "ocel_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "original",
                    "filtered"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "filtered",
              "title": "Ocel Version"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/OCELFilter"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Filter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/OCELFilter"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Setfilters"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ocels/ocel/delete": {
      "post": {
        "tags": [
          "ocels"
        ],
        "summary": "Delete an uploaded OCEL",
        "description": "Deletes the uploaded OCEL with the given `ocel_id`. This action is irreversible and removes the OCEL from the session.",
        "operationId": "deleteOcel",
        "parameters": [
          {
            "name": "ocel_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ocel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ocels/ocel/rename": {
      "post": {
        "tags": [
          "ocels"
        ],
        "summary": "Rename an uploaded OCEL",
        "description": "Renames the OCEL represented by the given `ApiOcel` object to `new_name`. This updates the display name used in the UI and metadata.",
        "operationId": "renameOcel",
        "parameters": [
          {
            "name": "new_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "New Name"
            }
          },
          {
            "name": "ocel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ocel Id"
            }
          },
          {
            "name": "ocel_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "original",
                    "filtered"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "filtered",
              "title": "Ocel Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ocels/objects/attributes": {
      "get": {
        "tags": [
          "ocels"
        ],
        "summary": "Get Object Attributes",
        "operationId": "objectAttributes",
        "parameters": [
          {
            "name": "ocel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ocel Id"
            }
          },
          {
            "name": "ocel_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "original",
                    "filtered"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "filtered",
              "title": "Ocel Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/IntegerAttribute"
                        },
                        {
                          "$ref": "#/components/schemas/FloatAttribute"
                        },
                        {
                          "$ref": "#/components/schemas/BooleanAttribute"
                        },
                        {
                          "$ref": "#/components/schemas/DateAttribute"
                        },
                        {
                          "$ref": "#/components/schemas/NominalAttribute"
                        }
                      ],
                      "discriminator": {
                        "propertyName": "type",
                        "mapping": {
                          "integer": "#/components/schemas/IntegerAttribute",
                          "float": "#/components/schemas/FloatAttribute",
                          "boolean": "#/components/schemas/BooleanAttribute",
                          "date": "#/components/schemas/DateAttribute",
                          "nominal": "#/components/schemas/NominalAttribute"
                        }
                      }
                    }
                  },
                  "title": "Response Objectattributes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ocels/events/attributes": {
      "get": {
        "tags": [
          "ocels"
        ],
        "summary": "Get Event Attributes",
        "operationId": "eventAttributes",
        "parameters": [
          {
            "name": "ocel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ocel Id"
            }
          },
          {
            "name": "ocel_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "original",
                    "filtered"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "filtered",
              "title": "Ocel Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "oneOf": [
                        {
                          "$ref": "#/components/schemas/IntegerAttribute"
                        },
                        {
                          "$ref": "#/components/schemas/FloatAttribute"
                        },
                        {
                          "$ref": "#/components/schemas/BooleanAttribute"
                        },
                        {
                          "$ref": "#/components/schemas/DateAttribute"
                        },
                        {
                          "$ref": "#/components/schemas/NominalAttribute"
                        }
                      ],
                      "discriminator": {
                        "propertyName": "type",
                        "mapping": {
                          "integer": "#/components/schemas/IntegerAttribute",
                          "float": "#/components/schemas/FloatAttribute",
                          "boolean": "#/components/schemas/BooleanAttribute",
                          "date": "#/components/schemas/DateAttribute",
                          "nominal": "#/components/schemas/NominalAttribute"
                        }
                      }
                    }
                  },
                  "title": "Response Eventattributes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ocels/events/counts": {
      "get": {
        "tags": [
          "ocels"
        ],
        "summary": "Get Event Counts",
        "operationId": "eventCounts",
        "parameters": [
          {
            "name": "ocel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ocel Id"
            }
          },
          {
            "name": "ocel_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "original",
                    "filtered"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "filtered",
              "title": "Ocel Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Eventcounts"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ocels/events/time": {
      "get": {
        "tags": [
          "ocels"
        ],
        "summary": "Get Time Info",
        "operationId": "timeInfo",
        "parameters": [
          {
            "name": "ocel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ocel Id"
            }
          },
          {
            "name": "ocel_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "original",
                    "filtered"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "filtered",
              "title": "Ocel Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entity_Time_Info"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ocels/objects/counts": {
      "get": {
        "tags": [
          "ocels"
        ],
        "summary": "Get Object Counts",
        "operationId": "objectCount",
        "parameters": [
          {
            "name": "ocel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ocel Id"
            }
          },
          {
            "name": "ocel_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "original",
                    "filtered"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "filtered",
              "title": "Ocel Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Objectcount"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ocels/relations/e2o": {
      "get": {
        "tags": [
          "ocels"
        ],
        "summary": "Get E2O",
        "operationId": "e2o",
        "parameters": [
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "source",
                    "target"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "source",
              "title": "Direction"
            }
          },
          {
            "name": "ocel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ocel Id"
            }
          },
          {
            "name": "ocel_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "original",
                    "filtered"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "filtered",
              "title": "Ocel Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RelationCountSummary"
                  },
                  "title": "Response E2O"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ocels/relations/o2o": {
      "get": {
        "tags": [
          "ocels"
        ],
        "summary": "Get Object Relations",
        "operationId": "o2o",
        "parameters": [
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "source",
                    "target"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "source",
              "title": "Direction"
            }
          },
          {
            "name": "ocel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ocel Id"
            }
          },
          {
            "name": "ocel_version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "original",
                    "filtered"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "filtered",
              "title": "Ocel Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RelationCountSummary"
                  },
                  "title": "Response O2O"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IntegerAttribute": {
        "properties": {
          "attribute": {
            "type": "string",
            "title": "Attribute"
          },
          "type": {
            "type": "string",
            "const": "integer",
            "title": "Type"
          },
          "min": {
            "type": "integer",
            "title": "Min"
          },
          "max": {
            "type": "integer",
            "title": "Max"
          }
        },
        "type": "object",
        "required": [
          "attribute",
          "type",
          "min",
          "max"
        ],
        "title": "IntegerAttribute"
      },
...
      "OCELFilter": {
        "properties": {
          "object_types": {
            "$ref": "#/components/schemas/ObjectTypeFilter"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventTypeFilter"
          },
          "time_range": {
            "$ref": "#/components/schemas/TimeFrameFilter"
          },
          "o2o_count": {
            "items": {
              "$ref": "#/components/schemas/O2OCountFilter"
            },
            "type": "array",
            "title": "O2O Count"
          },
          "e2o_count": {
            "items": {
              "$ref": "#/components/schemas/E2OCountFilter"
            },
            "type": "array",
            "title": "E2O Count"
          },
          "event_attributes": {
            "items": {
              "$ref": "#/components/schemas/EventAttributeFilter"
            },
            "type": "array",
            "title": "Event Attributes"
          },
          "object_attributes": {
            "items": {
              "$ref": "#/components/schemas/ObjectAttributeFilter"
            },
            "type": "array",
            "title": "Object Attributes"
          }
        },
        "type": "object",
        "title": "OCELFilter"
      },
...
      "OcelMetadata": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "extensions": {
            "items": {
              "$ref": "#/components/schemas/OCELExtensionDescription"
            },
            "type": "array",
            "title": "Extensions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at",
          "extensions"
        ],
        "title": "OcelMetadata"
      },
...
    }
  }
}